+++ src/core/default/commands/function/logs.py
@@ -13,7 +13,6 @@ from core.default.commands.function.utils import get_cloud_id_from_cdev_name
 
 
 class show_logs(BaseCommand):
-
     help = """
         Get the logs of a deployed lambda function
     """
@@ -33,22 +32,45 @@ class show_logs(BaseCommand):
             "--tail", action="store_true", help="show the tail of the logs"
         )
         parser.add_argument(
-            "--number",
-            action="store_true",
+            "--limit",
+            type=int,
+            nargs='?',
+            default=10000,
             help="number of events to show. Must be used with --tail.",
+        )
+        parser.add_argument(
+            "--query",
+            nargs='+',
+            help="retrieve information base on a cloudWatch query",
+        )
+        parser.add_argument(
+            "--start_time",
+            type=lambda d: datetime.datetime.strptime(d, '%Y-%m-%d-%H:%M:%S'),
+            help="set the start_time when use the query option, must be on the Y-m-d-H:M:S format",
+        )
+        parser.add_argument(
+            "--end_time",
+            type=lambda d: datetime.datetime.strptime(d, '%Y-%m-%d-%H:%M:%S'),
+            help="set the end_time when use the query option, must be on the Y-m-d-H:M:S format",
         )
 
     def command(self, *args, **kwargs) -> None:
+        (
+            component_name,
+            function_name,
+        ) = self.get_component_and_resource_from_qualified_name(
+            kwargs.get("function_name")
+        )
 
-        component_name, function_name = self.get_component_and_resource_from_qualified_name(kwargs.get("function_name"))
-
-        # tail_val = kwargs.get("tail")
-        # number_val = kwargs.get("number")
+        tail_val = kwargs.get("tail")
+        limit_val = kwargs.get("limit")
+        query_val = kwargs.get("query")
+        start_time_val = kwargs.get("start_time")
+        end_time_val = kwargs.get("end_time")
 
         cloud_name = get_cloud_id_from_cdev_name(component_name, function_name).split(
             ":"
         )[-1]
-
         if not cloud_name:
             self.stdout.write(
                 f"Could not find function {function_name} in component {component_name}"
@@ -63,25 +85,81 @@ class show_logs(BaseCommand):
             return
 
         cloud_watch_client = client("logs")
-        print(cloud_watch_group_name)
+        if query_val is None:
+            self.get_log_result(limit_val, tail_val, cloud_watch_client, cloud_watch_group_name)
+        else:
+            self.get_log_query_result(query_val, start_time_val, end_time_val,
+                                      cloud_watch_client, cloud_watch_group_name)
+
+    def get_log_result(self, limit_val, tail_val, cloud_watch_client, cloud_watch_group_name) -> None:
         log_streams_rv = cloud_watch_client.describe_log_streams(
             logGroupName=cloud_watch_group_name,
             orderBy="LastEventTime",
         )
-
         streams = _get_needed_streams(log_streams_rv.get("logStreams"))
-
         for stream in streams:
             response = cloud_watch_client.get_log_events(
                 logGroupName=cloud_watch_group_name,
                 logStreamName=stream,
-                startFromHead=True,
+                limit=limit_val,
+                startFromHead=False if tail_val else True,
+            )
+            if tail_val:
+                next_token = response.get("nextForwardToken")
+            else:
+                next_token = response.get("nextBackwardToken")
+            prev_token = ""
+            while next_token != prev_token:
+                for event in response.get("events"):
+                    self.stdout.write(
+                        f"{datetime.datetime.fromtimestamp(event.get('timestamp') / 1000).strftime('%Y-%m-%d %H:%M:%S')} - {event.get('message')} "
+                    )
+                response = cloud_watch_client.get_log_events(
+                    logGroupName=cloud_watch_group_name,
+                    logStreamName=stream,
+                    limit=limit_val,
+                    startFromHead=False if tail_val else True,
+                    nextToken=next_token
+                )
+                prev_token = next_token
+                if tail_val:
+                    next_token = response.get("nextForwardToken")
+                else:
+                    next_token = response.get("nextBackwardToken")
+
+    def get_log_query_result(self, query_val, start_time_val, end_time_val,
+                             cloud_watch_client, cloud_watch_group_name) -> None:
+        print(query_val)
+        print(start_time_val)
+        print(end_time_val)
+        print(cloud_watch_client)
+        print(cloud_watch_group_name)
+        if start_time_val is None:
+            start_time_val = datetime.datetime.today() - datetime.timedelta(weeks=52)
+        if end_time_val is None:
+            end_time_val = datetime.datetime.now()
+        query = ''
+        for item in query_val:
+            query = query + ' ' + str(item)
+        start_query_response = cloud_watch_client.start_query(
+            logGroupName=cloud_watch_group_name,
+            startTime=int(start_time_val.timestamp()),
+            endTime=int(end_time_val.timestamp()),
+            queryString=query,
+        )
+        query_id = start_query_response['queryId']
+        response = None
+
+        while response == None or response['status'] == 'Running':
+            self.stdout.write('Waiting for query to complete ...')
+            time.sleep(1)
+            response = cloud_watch_client.get_query_results(
+                queryId=query_id
             )
 
-            for event in response.get("events"):
-                self.stdout.write(
-                    f"{datetime.datetime.fromtimestamp(event.get('timestamp')/1000).strftime('%Y-%m-%d %H:%M:%S')} - {event.get('message')}"
-                )
+        for item in response['results']:
+            for item2 in item:
+                self.stdout.write(item2['field'] + ': ' + item2['value'])
 
 
 def _watch_log_group(group_name: str, stdout: OutputWrapper, args=None) -> None:
@@ -93,10 +171,18 @@ def _watch_log_group(group_name: str, stdout: OutputWrapper, args=None) -> None:
         _previous_refresh_time = int(
             (time.mktime(datetime.datetime.now().timetuple()) - 60) * 1000
         )
-        keep_watching = _read_from_streams(cloud_watch_client, group_name, stdout, events_hash, _previous_refresh_time)
+        keep_watching = _read_from_streams(
+            cloud_watch_client, group_name, stdout, events_hash, _previous_refresh_time
+        )
 
 
-def _read_from_streams(cloud_watch_client, group_name: str, stdout: OutputWrapper, events_hash, _previous_refresh_time) -> bool:
+def _read_from_streams(
+        cloud_watch_client,
+        group_name: str,
+        stdout: OutputWrapper,
+        events_hash,
+        _previous_refresh_time,
+) -> bool:
     try:
         log_streams_rv = cloud_watch_client.describe_log_streams(
             logGroupName=group_name,
@@ -126,7 +212,7 @@ def _read_from_streams(cloud_watch_client, group_name: str, stdout: OutputWrappe
 
                 if event_hash not in events_hash:
                     stdout.write(
-                        f"{datetime.datetime.fromtimestamp(event.get('timestamp')/1000).strftime('%Y-%m-%d %H:%M:%S')} - {event.get('message')}"
+                        f"{datetime.datetime.fromtimestamp(event.get('timestamp') / 1000).strftime('%Y-%m-%d %H:%M:%S')} - {event.get('message')}"
                     )
 
                 events_hash.add(event_hash)